requirements.txt
requests>=2.32.5
packaging>=25.0
loguru>=0.7.3
python-telegram-bot>=21.0
python-dotenv>=1.0.1
setuptools>=80.9.0
